datatype of price and total is changed to BigDecimal

+++ MyShopping/src/main/java/shop/dao/OrderDao.java
@@ -1,5 +1,6 @@
 package shop.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import shop.bean.Address;
@@ -9,7 +10,7 @@ import shop.bean.Order;
 
 public interface OrderDao {
 	
-	public int Addorder(String proname,String username,int quantity,double total);
+	public int Addorder(String proname,String username,int quantity,BigDecimal total);
 	public int addAddress(Address ad);
 	public List<Order> getOrders(String username);
 	public int cancelorder(int id);
